+++ packages/likec4/src/cli/codegen/webcomponent/handler.ts
@@ -1,5 +1,5 @@
 import { viteWebcomponentConfig } from '#vite/config-webcomponent'
-import { fromWorkspace } from '@likec4/language-services/node/without-mcp'
+import { fromWorkspace } from '@likec4/language-services/node'
 import { existsSync } from 'node:fs'
 import { copyFile, mkdir, rm, stat } from 'node:fs/promises'
 import { basename, dirname, extname, isAbsolute, relative, resolve } from 'node:path'