+++ packages/vscode-preview/protocol.ts
@@ -124,6 +124,7 @@ export const WebviewMsgs = {
     | { screen: 'view'; viewId: ViewId; projectId: ProjectId | undefined }
     | { screen: 'projects' }
   >,
+  OpenExternalUrl: { method: 'webview:open-external-url' } as NotificationType<{ url: string }>,
   UpdateMyTitle: { method: 'webview:update-my-title' } as NotificationType<{ title: string }>,
   EnhanceMeWithAI: { method: 'webview:enhance-me-ai' } as NotificationType<never>,
 }