+++ packages/vscode-preview/src/vscode.ts
@@ -55,6 +55,9 @@ export const ExtensionApi = {
   locate: (params: WebviewLocateReq) => {
     messenger.sendNotification(WebviewMsgs.Locate, HOST_EXTENSION, params)
   },
+  openExternalUrl: (url: string) => {
+    messenger.sendNotification(WebviewMsgs.OpenExternalUrl, HOST_EXTENSION, { url })
+  },
   updateTitle: (title: string) => {
     messenger.sendNotification(WebviewMsgs.UpdateMyTitle, HOST_EXTENSION, { title })
   },